extract log correlation to an external utility
